Cardano.Tracer.Configuration

Documentation

type Address = HowToConnect

data Certificate

Constructors

Certificate

Fields

certificateFile :: !FilePath

certificateKeyFile :: !FilePath

certificateChain :: !(Maybe [FilePath])

data HowToConnect

Constructors

LocalPipe !FilePath
Local pipe (UNIX or Windows).

RemoteSocket !Host !Port
Remote socket (host and port).

data Endpoint

Endpoint for internal services.

Constructors

Endpoint

Fields

epHost :: !String

epPort :: !Port

epForceSSL :: !(Maybe Bool)

Nothing (absent field) and `Just False' (False field) both disable SSL.

setEndpoint :: Endpoint -> Settings -> Settings

Endpoint {host, port} acting on Settings: setting host and port.

newtype FileOrMap

Constructors

FOM (Either FilePath (Map Text Text))

data LogFormat

Format of log files.

Constructors

ForHuman
For human (text)

ForMachine
For machine (JSON)

data LogMode

Logging mode.

Constructors

FileMode
Store items in log file.

JournalMode
Store items in Linux journal service.

data LoggingParams

Logging parameters.

Constructors

LoggingParams

Fields

logRoot :: !FilePath

Root directory where all subdirs with logs are created.

logMode :: !LogMode

Log mode.

logFormat :: !LogFormat

Log format.

data Network

Connection mode.

Constructors

AcceptAt !Address
Server mode: accepts connections.

ConnectTo !(NonEmpty Address)
Client mode: initiates connections.

data RotationParams

Parameters of rotation mechanism for logs.

Constructors

RotationParams

Fields

rpFrequencySecs :: !Word32

Rotation period, in seconds.

rpLogLimitBytes :: !Word64

Max size of log file in bytes.

rpMaxAgeMinutes :: !Word64

Max age of log file in minutes.

rpKeepFilesNum :: !Word32

Number of log files to keep in any case.

data TracerConfig

Tracer configuration.

Constructors

TracerConfig

Fields

networkMagic :: !Word32

Network magic from genesis the node is launched with.

network :: !Network

How cardano-tracer will be connected to node(s).

loRequestNum :: !(Maybe Word16)

How many TraceObjects will be asked in each request.

ekgRequestFreq :: !(Maybe Pico)

How often to request for EKG-metrics, in seconds.

hasEKG :: !(Maybe Endpoint)

Endpoint for EKG web-page.

hasPrometheus :: !(Maybe Endpoint)

Endpoint for Prometheus web-page.

hasRTView :: !(Maybe Endpoint)

Endpoint for RTView web-page.

tlsCertificate :: !(Maybe Certificate)

hasForwarding :: !(Maybe (Network, Maybe [[Text]], TraceOptionForwarder))

Socket for tracer's to reforward on. Second member of the triplet is the list of prefixes to reforward. Third member of the triplet is the forwarder config.

logging :: !(NonEmpty LoggingParams)

Logging parameters.

rotation :: !(Maybe RotationParams)

Rotation parameters.

verbosity :: !(Maybe Verbosity)

Verbosity of the tracer itself.

metricsNoSuffix :: !(Maybe Bool)

Prometheus ONLY: Dropping metrics name suffixes (like "_int") increases similarity with old system names - if desired; default: False

metricsHelp :: !(Maybe FileOrMap)

Prometheus ONLY: JSON file or object containing a key-value map "metric name -> help text" for "# HELP " annotations

resourceFreq :: !(Maybe Int)

Frequency (1/millisecond) for gathering resource data.

ekgRequestFull :: !(Maybe Bool)

Request full set of metrics always, vs. deltas only (safer, but more overhead); default: False

prometheusLabels :: !(Maybe (Map Text Text))

A common label set for all Prometheus scrape targets (only used in Prometheus HTTP service discovery)

data Verbosity

Tracer's verbosity.

Constructors

Minimum
Display minimum of messages.

ErrorsOnly
Display errors only.

Maximum
Display all the messages (protocols tracing, errors).

readTracerConfig :: FilePath -> IO TracerConfig

Read the tracer's configuration file.