Cardano.Tracer.Handlers.Metrics.Monitoring

Documentation

runMonitoringServer

Arguments

:: TracerEnv

-> Endpoint
(web page with list of connected nodes, EKG web page).

-> IO RouteDictionary

-> IO ()

ekg package allows to run only one EKG server, to display only one web page for particular EKG.Store. Since 'cardano-tracer' can be connected to any number of nodes, we display their list on the first web page (the first Endpoint) as a list of hrefs. After clicking on the particular href, the user will be redirected to the monitoring web page (the second Endpoint) built by ekg package. This page will display the metrics received from that node.

If the user returns to the first web page and clicks to another node's href, the EKG server will be restarted and the monitoring page will display the metrics received from that node.