Cardano.Tracer.Handlers.Metrics.Prometheus

Documentation

runPrometheusServer :: TracerEnv -> Endpoint -> IO RouteDictionary -> IO ()

Runs a simple HTTP server that listens on endpoint.

At the root, it lists the connected nodes, either as HTML or JSON, depending on the request's 'Accept: ' header.

Routing is dynamic, depending on the connected nodes. A valid URL is derived from the nodeName configured for the connecting node. E.g. a node name of `127.0.0.1:30004` will result in the route `/12700130004` which renders that node's Prometheus / OpenMetrics text exposition:

# TYPE Mem_resident_int gauge # HELP Mem_resident_int Kernel-reported RSS (resident set size) Mem_resident_int 103792640 # TYPE RTS_gcMajorNum_int gauge # HELP RTS_gcMajorNum_int Major GCs RTS_gcMajorNum_int 4 # TYPE rts_gc_num_bytes_usage_samples counter rts_gc_num_bytes_usage_samples 4 # TYPE remainingKESPeriods_int gauge remainingKESPeriods_int 62 # TYPE rts_gc_bytes_copied counter rts_gc_bytes_copied 17114384 # TYPE nodeCannotForge_int gauge # HELP nodeCannotForge_int How many times was this node unable to forge [a block]?

The `/targets` path can be used for Prometheus HTTP service discovery. This lets Prometheus dynamically discover all connected nodes, and scrape their metrics. Below is a minimal example of a corresponding job definition that goes into the `prometheus.yml` configuration:

job_name: "cardano-tracer"

http_sd_configs: - url: 'http://127.0.0.1:3200/targets' # <-- Your cardano-tracer's real hostname:prometheus port

Each target will have a label "node_name" which corresponds to the TraceOptionNodeName setting in the node config.

In cardano-tracer's config, you can optionally provide additional labels to be attached to *all* targets (default is no additional labels): "prometheusLabels": { "labelname": "labelvalue", ... }