Cardano.Tracer.Handlers.RTView.State.Displayed

Documentation

type DisplayedElements = TVar (Map NodeId DisplayedForNode)

type PageReloadedFlag = TVar Bool

If the user reloaded the web-page, after DOM re-rendering, we have to restore displayed state of all elements that they have _before_ page's reload.

cleanupDisplayedValues :: DisplayedElements -> IO ()

getDisplayedValue :: DisplayedElements -> NodeId -> ElementId -> IO (Maybe ElementValue)

getDisplayedValuePure :: Map NodeId DisplayedForNode -> NodeId -> ElementId -> Maybe ElementValue

initDisplayedElements :: IO DisplayedElements

initPageReloadFlag :: IO PageReloadedFlag

pageWasReload :: PageReloadedFlag -> IO ()

pageWasNotReload :: PageReloadedFlag -> IO ()

saveDisplayedValue :: DisplayedElements -> NodeId -> ElementId -> ElementValue -> IO ()

updateDisplayedElements :: DisplayedElements -> Set NodeId -> IO ()

Makes displayedElements up-to-date with connected. There are 3 cases:

displayedElements contains all node ids from connected - just keep it.

displayedElements contains some node ids that is not presented in connected - remove them, these ids correspond to disconnected nodes.

displayedElements doesn't contain some node ids that presented in connected - add them, these ids correspond to newly connected nodes.

2-nd and 3-d cases can coexist.