Cardano.Tracer.Handlers.RTView.State.Historical

Documentation

newtype BlockchainHistory

Constructors

ChainHistory History

data DataName

Historical points for particular data.

Constructors

CPUData

MemoryData

GCMajorNumData

GCMinorNumData

GCLiveMemoryData

CPUTimeGCData

CPUTimeAppData

ThreadsNumData

ChainDensityData

SlotNumData

BlockNumData

SlotInEpochData

EpochData

NodeCannotForgeData

ForgedSlotLastData

NodeIsLeaderData

NodeIsNotLeaderData

ForgedInvalidSlotLastData

AdoptedSlotLastData

NotAdoptedSlotLastData

AboutToLeadSlotLastData

CouldNotForgeSlotLastData

TxsProcessedNumData

MempoolBytesData

TxsInMempoolData

type History = TVar (Map NodeId HistoricalData)

type HistoricalData = Map DataName HistoricalPoints

type HistoricalPoint = (POSIXTime, ValueH)

type HistoricalPoints = Set HistoricalPoint

type POSIXTime = Word64

A lot of information received from the node is useful as historical data. It means that such an information should be displayed on time charts, where X axis is a time in UTC. An example: resource metrics, chain information, tx information, etc.

newtype ResourcesHistory

Constructors

ResHistory History

newtype TransactionsHistory

Constructors

TXHistory History

data ValueH

Constructors

ValueD Double

ValueI Int

addHistoricalData :: History -> NodeId -> UTCTime -> DataName -> ValueH -> IO ()

getHistoricalData :: History -> NodeId -> DataName -> IO [(POSIXTime, ValueH)]

initBlockchainHistory :: IO BlockchainHistory

initResourcesHistory :: IO ResourcesHistory

initTransactionsHistory :: IO TransactionsHistory

readValueI :: Monad m => Text -> (ValueH -> m ()) -> m ()

readValueD :: Monad m => Text -> (ValueH -> m ()) -> m ()