Cardano.Tracer.Handlers.RTView.State.Last

Documentation

data LastResourcesForNode

We have to store last received metric to be able to calculate the value based on next received metric. For example, calculate "CPU usage percent" based on received CPU ticks. That value will be stored in corresponding HistoricalData for rendering on the corresponding chart.

Constructors

LastResourcesForNode

Fields

cpuLastTicks :: !Int

cpuLastNS :: !Word64

cpuGCLastTicks :: !Int

cpuGCLastNS :: !Word64

cpuAppLastTicks :: !Int

cpuAppLastNS :: !Word64

type LastResources = TVar (Map NodeId LastResourcesForNode)

addNullResources :: LastResources -> NodeId -> IO ()

For the first update only.

initLastResources :: IO LastResources

updateLastResources :: LastResources -> NodeId -> (LastResourcesForNode -> LastResourcesForNode) -> IO ()