Cardano.Tracer.Handlers.RTView.UI.Logs

Documentation

restoreLogsLiveViewFont :: TracerEnv -> UI ()

When the user opened the logs live view window, the previous font's size should be restored from the disk.

saveLogsLiveViewFont :: TracerEnv -> String -> UI ()

Every time when the user changed the font's size in logs live view window, it should be saved on the file for next sessions, both after web-page reload and 'cardano-tracer' restart.