Cardano.Tracer.Handlers.RTView.UI.Types

Documentation

data ChartId

Constructors

CPUChart

MemoryChart

GCMajorNumChart

GCMinorNumChart

GCLiveMemoryChart

CPUTimeGCChart

CPUTimeAppChart

ThreadsNumChart

ChainDensityChart

SlotNumChart

BlockNumChart

SlotInEpochChart

EpochChart

NodeCannotForgeChart

ForgedSlotLastChart

NodeIsLeaderChart

NodeIsNotLeaderChart

ForgedInvalidSlotLastChart

AdoptedSlotLastChart

NotAdoptedSlotLastChart

AboutToLeadSlotLastChart

CouldNotForgeSlotLastChart

TxsProcessedNumChart

MempoolBytesChart

TxsInMempoolChart

data ChartSelectId

Constructors

TimeRangeSelect

UpdatePeriodSelect

data ChartSettings

Constructors

ChartSettings

Fields

csTimeRangeInS :: !Int

csUpdatePeriodInS :: !Int

type ChartsSettings = [(ChartId, ChartSettings)]

newtype Color

Constructors

Color String

type Colors = TBQueue Color

type DatasetsIndices = TVar (Map NodeName Index)

After the node is connected, we have to add a new dataset to all historical charts. The metrics received from this node will be added in these datasets. Since each dataset has its index, we need a map 'NodeName -> ix', where ix is an index of a dataset in _each_ chart. We use node's name instead of node's id, because we want to keep the same color for the same node.

newtype Index

Constructors

Index Word16

type WebPageStatus = TVar Bool

It's True if the web page is opened, False otherwise.