Cardano.Tracer.Handlers.RTView.Update.Historical

Documentation

backupAllHistory :: TracerEnv -> IO ()

backupSpecificHistory :: TracerEnv -> History -> [NodeId] -> DataName -> IO ()

Backup specific history after these points were pushed to corresponding JS-chart. After cleanup we can safely remove them from the memory.

getAllHistoryFromBackup :: TracerEnv -> DataName -> IO [(NodeId, [HistoricalPoint])]

getLastHistoryFromBackups :: TracerEnv -> Set NodeId -> IO [(NodeId, [(DataName, [HistoricalPoint])])]

getLastHistoryFromBackupsAll :: TracerEnv -> IO [(NodeId, [(DataName, [HistoricalPoint])])]

runHistoricalBackup :: TracerEnv -> IO ()

If RTView's web page is opened, historical backup is performing by UI-code, in this case we should skip backup.

runHistoricalUpdater :: TracerEnv -> LastResources -> IO ()

A lot of information received from the node is useful as historical data. It means that such an information should be displayed on time charts, where X axis is a time in UTC. An example: resource metrics, chain information, tx information, etc.

This information is extracted both from TraceObjects and Metrics and then it will be saved as chart coords '[(ts, v)]', where ts is a timestamp and v is a value. Later, when the user will open RTView web-page, this saved data will be used to render historical charts.

It allows to collect historical data even when RTView web-page is closed.