Cardano.Tracer.Handlers.ReForwarder

Description

This module initializes a reforwarding service for use by cardano-tracer. It could [re-] serve the three miniprotocols on a new local socket. Currently, - it reforwards trace messages to the new socket, optionally filtering trace messages. - it does not (currently) reforward EKG to the new socket. - it creates a new Datapoint store, but the datapoint store is empty.

Documentation

initReForwarder :: TracerConfig -> Trace IO TracerTrace -> IO ([TraceObject] -> IO (), Trace IO DataPoint)

Initialize the reforwarding service if configured to be active. Returns - the function by which logging sources report their log messages - the DataPoint tracer (for the data point store associated with the forwarding trace server).