Cardano.Tracer.Handlers.Utils

Documentation

readInt :: Text -> Int -> Int

normalizeNamespace :: [Text] -> Text

askDataPoint :: FromJSON a => DataPointRequestors -> Lock -> NodeId -> DataPointName -> IO (Maybe a)

There is a different information the node can provide us by explicit request. This is a structured data about internal state of the node (for example, its basic information like version, protocol, commit hash, start time, etc).

Such a structured data is provided as a DataPoint. When it's receved, it's technically a lazy bytestring that is a result of ToJSON-encoding on the forwarder's side. Here we can decode it to particular Haskell type provided by the node.

utc2s :: UTCTime -> Word64

Converts a timestamp to seconds since Unix epoch.

utc2ns :: UTCTime -> Word64

Converts a timestamp to nanoseconds since Unix epoch.

s2utc :: Word64 -> UTCTime

nullTime :: UTCTime