Cardano.Tracer.MetaTrace

Documentation

mkTimeseriesTracer :: Trace IO FormattedMessage -> IO (Trace IO TimeseriesTrace)

mkTraceBundle :: SeverityF -> IO TraceBundle

mkTracerTracer :: Trace IO FormattedMessage -> SeverityF -> IO (Trace IO TracerTrace)

stderrShowTracer :: Show a => Trace IO a

data TraceBundle

A bundle of domain-split tracers used in the application.

Constructors

TraceBundle

Fields

assorted :: !(Trace IO TracerTrace)

A tracer used to trace all kinds of things happening in the application.

timeseries :: !(Trace IO TimeseriesTrace)

A tracer that has to do only with timeseries storingqueryingpruning etc.

data TracerTrace

Constructors

TracerBuildInfo

TracerParamsAre

Fields

ttConfigPath :: FilePath

ttStateDir :: Maybe FilePath

ttMinLogSeverity :: Maybe SeverityS

TracerConfigIs

Fields

ttConfig :: TracerConfig

TracerInitStarted

TracerInitDone

TracerAddNewNodeIdMapping

Fields

ttBimapping :: !(NodeId, NodeName)

TracerStartedLogRotator

TracerStartedPrometheus

Fields

ttPrometheusEndpoint :: Endpoint

TracerStartedTimeseries

Fields

ttTimeseriesEndpoint :: Endpoint

TracerStartedMonitoring

Fields

ttMonitoringEndpoint :: Endpoint

ttMonitoringType :: Text

TracerStartedAcceptors

Fields

ttAcceptorsAddr :: Network

TracerStartedReforwarder

TracerSockListen

Fields

ttListenAt :: FilePath

TracerSockIncoming

Fields

ttConnectionIncomingAt :: FilePath

ttAddr :: Text

TracerSockConnecting

Fields

ttConnectingTo :: FilePath

TracerSockConnected

Fields

ttConnectedTo :: FilePath

TracerShutdownInitiated

TracerShutdownComplete

TracerError

Fields

ttError :: Text

TracerMissingCertificate

Fields

ttMissingCertificateEndpoint :: !Endpoint

TracerResource

Fields

ttResource :: ResourceStats

TracerForwardingInterrupted

Fields

ttConnection :: HowToConnect

ttMessage :: String

data Trace (m :: Type -> Type) a

newtype SeverityF

Constructors

SeverityF (Maybe SeverityS)

data SeverityS

Constructors

Debug

Info

Notice

Warning

Error

Critical

Alert

Emergency

traceWith :: Monad m => Trace m a -> a -> m ()