Cardano.Tracer.Types

Documentation

type AcceptedMetrics = TVar (Map NodeId MetricsStores)

We have to create EKG.Store and MetricsLocalStore to keep all the metrics accepted from the node.

type ConnectedNodes = TVar (Set NodeId)

This set contains ids of currently connected nodes. When the node is connected, its id will be added to this set, and when it is disconnected, its id will be deleted from this set. So, ConnectedNodes is used as a "source of truth" about currently connected nodes.

type ConnectedNodesNames = TVar (Bimap NodeId NodeName)

We have to map node's id to its name (received with NodeInfo).

type DataPointRequestors = TVar (Map NodeId DataPointRequestor)

We have to store DataPointRequestors to be able to ask particular node for some DataPoints.

type MetricsStores = (Store, TVar MetricsLocalStore)

newtype NodeId

Unique identifier of connected node, based on remoteAddress from ConnectionId, please see 'ouroboros-network'.

Constructors

NodeId Text

type NodeName = Text

type ProtocolsBrake = TVar Bool

The flag we use to stop the protocols from their acceptor's side.

newtype Registry a b

Constructors

Registry

Fields

getRegistry :: MVar (Map a b)

type HandleRegistry = Registry HandleRegistryKey (Handle, FilePath)

type HandleRegistryKey = (NodeName, LoggingParams)