Cardano.Tracer.Utils

Documentation

applyBrake :: ProtocolsBrake -> IO ()

Stop the protocols. As a result, MsgDone will be sent and interaction between acceptor's part and forwarder's part will be finished.

askNodeId :: TracerEnv -> NodeName -> IO (Maybe NodeId)

askNodeName :: TracerEnv -> NodeId -> IO NodeName

askNodeNameRaw :: Trace IO TracerTrace -> ConnectedNodesNames -> DataPointRequestors -> Lock -> NodeId -> IO NodeName

beforeProgramStops :: IO () -> IO ()

If 'cardano-tracer' process is going to die (by receiving some system signal), we want to do something before it stops.

connIdToNodeId :: Show addr => ConnectionId addr -> NodeId

initAcceptedMetrics :: IO AcceptedMetrics

initConnectedNodes :: IO ConnectedNodes

initConnectedNodesNames :: IO ConnectedNodesNames

initDataPointRequestors :: IO DataPointRequestors

initProtocolsBrake :: IO ProtocolsBrake

forMM :: (Traversable t, Monad m) => m (t a) -> (a -> m b) -> m (t b)

forMM_ :: (Foldable t, Monad m) => m (t a) -> (a -> m ()) -> m ()

nl :: Text

showProblemIfAny

Arguments

:: Maybe Verbosity
Tracer's verbosity.

-> Trace IO TracerTrace
Trace an error with that tracer iff not at minimum verbosity

-> IO ()
An IO action that can throw an exception.

-> IO ()

memberRegistry :: Ord a => a -> Registry a b -> IO Bool

showRegistry :: (Show a, Show b) => Registry a b -> IO ()

newRegistry :: IO (Registry a b)

lookupRegistry :: (Ord a, Ord b) => a -> b -> Registry (a, b) c -> IO (Maybe c)

elemsRegistry :: Registry a b -> IO [b]

clearRegistry :: HandleRegistry -> IO ()

modifyRegistry_ :: Registry a b -> (Map a b -> IO (Map a b)) -> IO ()

readRegistry :: Registry a b -> IO (Map a b)

getProcessId :: IO Word32

sanitizeNodeName :: Text -> Text

sequenceConcurrently_ :: Traversable t => t (IO a) -> IO ()

newtype NodeStateWrapper

This is to avoid creating a dependency on `cardano-node's NodeState.

Before: Pattern matching on NodeState to access a single Double:

  NodeAddBlock (AddedToCurrentChain _ _ syncPct) -> setSyncProgress nodeId syncPct
  _ -> return ()

Now: We pattern match on a newtype wrapper for a Double and parse it from a JSON object if it matches the serialization of `NodeAddBlock (AddedToCurrentChain _ _ syncPct)'.

  (NodeStateWrapper syncPct) -> setSyncProgress nodeId syncPct

Constructors

NodeStateWrapper

Fields

getNodeStateWrapper :: Double