Cardano.Tracer.Configuration

Documentation

newtype Address

Only local socket is supported, to avoid unauthorized connections.

Constructors

LocalSocket FilePath

data Endpoint

Endpoint for internal services.

Constructors

Endpoint

Fields

epHost :: !String

epPort :: !Word16

data LogFormat

Format of log files.

Constructors

ForHuman
For human (text)

ForMachine
For machine (JSON)

data LogMode

Logging mode.

Constructors

FileMode
Store items in log file.

JournalMode
Store items in Linux journal service.

data LoggingParams

Logging parameters.

Constructors

LoggingParams

Fields

logRoot :: !FilePath

Root directory where all subdirs with logs are created.

logMode :: !LogMode

Log mode.

logFormat :: !LogFormat

Log format.

data Network

Connection mode.

Constructors

AcceptAt !Address
Server mode: accepts connections.

ConnectTo !(NonEmpty Address)
Client mode: initiates connections.

data RotationParams

Parameters of rotation mechanism for logs.

Constructors

RotationParams

Fields

rpFrequencySecs :: !Word32

Rotation period, in seconds.

rpLogLimitBytes :: !Word64

Max size of log file in bytes.

rpMaxAgeMinutes :: !Word64

Max age of log file in minutes.

rpKeepFilesNum :: !Word32

Number of log files to keep in any case.

data TracerConfig

Tracer configuration.

Constructors

TracerConfig

Fields

networkMagic :: !Word32

Network magic from genesis the node is launched with.

network :: !Network

How cardano-tracer will be connected to node(s).

loRequestNum :: !(Maybe Word16)

How many TraceObjects will be asked in each request.

ekgRequestFreq :: !(Maybe Pico)

How often to request for EKG-metrics, in seconds.

hasEKG :: !(Maybe (Endpoint, Endpoint))

Endpoint for EKG web-page (list of nodes, monitoring).

hasPrometheus :: !(Maybe Endpoint)

Endpoint for Prometheus web-page.

hasRTView :: !(Maybe Endpoint)

Endpoint for RTView web-page. | Socket for tracer's to reforward on. Second member of the triplet is the list of prefixes to reforward. Third member of the triplet is the forwarder config.

hasForwarding :: !(Maybe (Network, Maybe [[Text]], TraceOptionForwarder))

logging :: !(NonEmpty LoggingParams)

Logging parameters.

rotation :: !(Maybe RotationParams)

Rotation parameters.

verbosity :: !(Maybe Verbosity)

Verbosity of the tracer itself.

metricsComp :: !(Maybe (Map Text Text))

Metrics compatibility map from metrics name to metrics name

resourceFreq :: !(Maybe Int)

Frequency (1/millisecond) for gathering resource data.

data Verbosity

Tracer's verbosity.

Constructors

Minimum
Display minimum of messages.

ErrorsOnly
Display errors only.

Maximum
Display all the messages (protocols tracing, errors).

readTracerConfig :: FilePath -> IO TracerConfig

Read the tracer's configuration file.