Cardano.Tracer.Handlers.Metrics.Prometheus

Documentation

runPrometheusServer :: TracerEnv -> Endpoint -> IO ()

Runs simple HTTP server that listens host and port and returns the list of currently connected nodes in such a format:

relay-1

relay-2

core-1

where 'relay-1', 'relay-2' and 'core-1' are nodes' names.

Each of list items is a href. By clicking on it, the user will be redirected to the page with the list of metrics received from that node, in such a format:

rts_gc_par_tot_bytes_copied 0 rts_gc_num_gcs 17 rts_gc_max_bytes_slop 15888 rts_gc_bytes_copied 165952 ekg_server_timestamp_ms 1639569439623